# contracts/openapi.yaml
{
    "openapi": "3.1.0",
    "info": {
        "title": "Zero to One API",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/projects/intake": {
            "post": {
                "summary": "Submit Idea",
                "operationId": "submit_idea_api_v1_projects_intake_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IntakeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClarityResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/intake/answers": {
            "post": {
                "summary": "Submit Answers",
                "operationId": "submit_answers_api_v1_projects_intake_answers_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClarityAnswersRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClarityResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/goals": {
            "post": {
                "summary": "Get Goals",
                "operationId": "get_goals_api_v1_projects_goals_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GoalsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects": {
            "get": {
                "summary": "List Projects",
                "operationId": "list_projects_api_v1_projects_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    },
                                    "type": "array",
                                    "title": "Response List Projects Api V1 Projects Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Project",
                "operationId": "create_project_api_v1_projects_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}": {
            "patch": {
                "summary": "Update Project",
                "operationId": "update_project_api_v1_projects__project_id__patch",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get Project",
                "operationId": "get_project_api_v1_projects__project_id__get",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/steps/{step_id}": {
            "patch": {
                "summary": "Update Step",
                "operationId": "update_step_api_v1_steps__step_id__patch",
                "parameters": [
                    {
                        "name": "step_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Step Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStepRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Step"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/milestones/{milestone_id}": {
            "patch": {
                "summary": "Update Milestone",
                "operationId": "update_milestone_api_v1_milestones__milestone_id__patch",
                "parameters": [
                    {
                        "name": "milestone_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Milestone Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMilestoneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Milestone"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/steps/{step_id}/tasks": {
            "get": {
                "summary": "Get Tasks",
                "operationId": "get_tasks_api_v1_steps__step_id__tasks_get",
                "parameters": [
                    {
                        "name": "step_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Step Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    },
                                    "title": "Response Get Tasks Api V1 Steps  Step Id  Tasks Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/{task_id}": {
            "patch": {
                "summary": "Update Task",
                "operationId": "update_task_api_v1_tasks__task_id__patch",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Task Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chat/sessions": {
            "post": {
                "summary": "Open Session",
                "operationId": "open_session_api_v1_chat_sessions_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OpenSessionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatSession"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chat/sessions/{session_id}/messages": {
            "post": {
                "summary": "Send Message",
                "operationId": "send_message_api_v1_chat_sessions__session_id__messages_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Session Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ClarityAnswersRequest": {
                "properties": {
                    "idea": {
                        "type": "string",
                        "title": "Idea"
                    },
                    "answers": {
                        "items": {
                            "$ref": "#/components/schemas/QAPair"
                        },
                        "type": "array",
                        "title": "Answers"
                    }
                },
                "type": "object",
                "required": [
                    "idea",
                    "answers"
                ],
                "title": "ClarityAnswersRequest"
            },
            "ClarityResult": {
                "properties": {
                    "clarity_score": {
                        "type": "number",
                        "title": "Clarity Score"
                    },
                    "needs_clarification": {
                        "type": "boolean",
                        "title": "Needs Clarification"
                    },
                    "clarifying_questions": {
                        "items": {
                            "$ref": "#/components/schemas/ClarifyingQuestion"
                        },
                        "type": "array",
                        "title": "Clarifying Questions"
                    },
                    "enriched_idea": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Enriched Idea"
                    }
                },
                "type": "object",
                "required": [
                    "clarity_score",
                    "needs_clarification",
                    "clarifying_questions"
                ],
                "title": "ClarityResult"
            },
...
            "GoalsRequest": {
                "properties": {
                    "category": {
                        "type": "string",
                        "title": "Category"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "idea": {
                        "type": "string",
                        "title": "Idea"
                    }
                },
                "type": "object",
                "required": [
                    "category",
                    "description",
                    "idea"
                ],
                "title": "GoalsRequest"
            },
            "GoalsResponse": {
                "properties": {
                    "goals": {
                        "items": {
                            "$ref": "#/components/schemas/Goal"
                        },
                        "type": "array",
                        "title": "Goals"
                    }
                },
                "type": "object",
                "required": [
                    "goals"
                ],
                "title": "GoalsResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IntakeRequest": {
                "properties": {
                    "category": {
                        "type": "string",
                        "title": "Category"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "idea": {
                        "type": "string",
                        "title": "Idea"
                    }
                },
                "type": "object",
                "required": [
                    "category",
                    "description",
                    "idea"
                ],
                "title": "IntakeRequest"
            },
...
            "SendMessageRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "SendMessageRequest"
            },
...
        }
    }
}